cleanup Gruntfile.js to be more maintainable
