+++ package.json
@@ -82,6 +82,8 @@
         "require-dir": "~0.1.0",
         "should": "~2.1.1",
         "sinon": "~1.7.3",
-        "supertest": "~0.8.2"
+        "supertest": "~0.8.2",
+        "load-grunt-tasks": "~0.4.0",
+        "load-grunt-config": "~0.7.1"
     }
 }
